requirements.txt
arrow==1.3.0
asgiref==3.7.2
blessed==1.20.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
cryptography==42.0.5
Django==4.2.11
django-crispy-forms==2.1
django-login-required-middleware==0.9.0
django-picklefield==3.1
django-q==1.3.9
django-recaptcha==4.0.0
django-stronghold==0.4.0
et-xmlfile==1.1.0
idna==3.6
openpyxl==3.1.2
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pycparser==2.21
pyOpenSSL==24.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==3.5.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
urllib3==2.2.1
wcwidth==0.2.13
